Cambio

+++ 34/34k.c
@@ -3,11 +3,11 @@
 #include <math.h>
 
 int main(int argc, char *argv[]){
-    int u1 = atoi(argv[1]);
-    int u2 = atoi(argv[2]);
-    int v1 = atoi(argv[3]);
-    int v2 = atoi(argv[4]);
-    float producto = u1*v1 + u2*v2;
-    printf("%f\n", producto);
+    float u1 = atof(argv[1]);
+    float u2 = atof(argv[2]);
+    float v1 = atof(argv[3]);
+    float v2 = atof(argv[4]);
+    float producto = (u1 * v1) + (u2 * v2);
+    printf("%.4f\n", producto);
     return 0;
 }
