+++ 34/34o.c
@@ -2,9 +2,9 @@
 #include <stdlib.h>
 
 int main(int argc, char *argv[]){
-    int horas = atoi(argv[1]);
-    int salario = atoi(argv[2]);
+    float horas = atof(argv[1]);
+    float salario = atof(argv[2]);
     float salxhora = horas*salario;
-    printf("%.2f\n", salxhora);
+    printf("%.4f\n", salxhora);
     return 0;
 }
